// projects/ontimize-web-ngx-extra-components/src/lib/components/o-data-view/list-config.types.ts
export interface ListConfig {
  controls?: string;
  detailFormRoute?: string;
  detailMode?: string;
  enabled?: string;
  pageSizeOptions?: string;
  quickFilter?: string;
  quickFilterPlaceholder?: string;
  recursiveDetail?: string;
  title?: string;
  visible?: string;
  detailButtonInRow?: string;
  detailButtonInRowIcon?: string;
  editButtonInRow?: string;
  editButtonInRowIcon?: string;
  editFormRoute?: string;
  insertButton?: string;
  insertFormRoute?: string;
  quickFilterAppearance?: string;
  recursiveEdit?: string;
  recursiveInsert?: string;
  rowHeight?: string;
  deleteButton?: 'no' | 'false' | 'yes' | 'true';
  insertButtonFloatable?: 'no' | 'false' | 'yes' | 'true';
  insertButtonPosition?: 'botton' | 'top';
  keysSqlTypes?: string;
  paginationControls?: 'no' | 'false' | 'yes' | 'true';
  quickFilterColumns?: string;
  refreshButton?: 'no' | 'false' | 'yes' | 'true';
  selectable?: 'no' | 'false' | 'yes' | 'true';
  showButtonsText?: 'no' | 'false' | 'yes' | 'true';
  sortColumns?: string;
}